Include upvote, downvote, delete buttons in post detail page
